developing...
Wrap the HK code in a function and add a main python file. I made a
simple sanity check to the HK code. I think more checks are needed but
I have no simple idea for now. And, the speed of the code seems not
that good. Also some minor changes to make PyCharm happy.